# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dientes Fijos Medellín CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /whatsapp/bot-toggle:
    post:
      operationId: toggleWhatsappBot
      tags: [whatsapp]
      summary: Enable or disable global AI bot
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  botEnabled:
                    type: boolean
...
  /whatsapp/send:
    post:
      operationId: sendWhatsappMessage
      tags: [whatsapp]
      summary: Send manual WhatsApp text message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                message:
                  type: string
              required: [phone, message]
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /clinical/evolution/{patientId}:
    get:
      operationId: listEvolutionNotes
      tags: [clinical]
      summary: List evolution notes for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /clinical/evolution:
    post:
      operationId: createEvolutionNote
      tags: [clinical]
      summary: Create evolution note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvolutionNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvolutionNote"
  /clinical/quotations:
    get:
      operationId: listQuotations
      tags: [clinical]
      summary: List quotations
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [clinical]
      summary: Create quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /clinical/quotations/{id}:
    patch:
      operationId: updateQuotation
      tags: [clinical]
      summary: Update quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuotationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
    delete:
      operationId: deleteQuotation
      tags: [clinical]
      summary: Delete quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        content:
          type: string
        sender:
          type: string
          enum: [patient, agent, ai]
        sentAt:
          type: string
          format: date-time
        read:
          type: boolean
        sentToWhatsApp:
          type: boolean
          description: True when agent/AI message was delivered via WhatsApp
      required: [id, conversationId, content, sender, sentAt, read]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    EvolutionNote:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        content:
          type: string
        doctorName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, content, createdAt]
    CreateEvolutionNoteBody:
      type: object
      properties:
        patientId:
          type: integer
        content:
          type: string
        doctorName:
          type: string
      required: [patientId, content]
    QuotationItem:
      type: object
      properties:
        service:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required: [service, price]
    Quotation:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, items, total, status, createdAt]
    CreateQuotationBody:
      type: object
      properties:
        patientId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        total:
          type: number
        status:
          type: string
        sendToWhatsApp:
          type: boolean
      required: [patientId, items, total]
    UpdateQuotationBody:
      type: object
      properties:
        patientId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        total:
          type: number
        status:
          type: string
        sendToWhatsApp:
          type: boolean
...