# lib/api-zod/src/generated/types/createQuotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dientes Fijos Medellín CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { QuotationItem } from "./quotationItem";

export interface CreateQuotationBody {
  patientId: number;
  items: QuotationItem[];
  total: number;
  status?: string;
  sendToWhatsApp?: boolean;
}